:truck: Move database and state code to `database` module
