+++ eadk_discord/database/__init__.py
@@ -1,0 +1,1 @@
+from .database import Database as Database
